adding typescript paths to important dirs, moving unsubscribe to shared
