+++ src/app/views/foods-list/foods-list.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { FormControl } from '@angular/forms';
 import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { UnsubscribeComponent } from '@app/shared/utils/unsubscribe';
 import { debounceTime, filter, startWith, takeUntil } from 'rxjs';
-import { UnsubscribeComponent } from 'src/app/components/unsubscribe';
 import { Food } from 'src/app/interfaces/food.interface';
 import { DialogComponent } from '../../components/dialog/dialog.component';
 import { DISPLAYED_COLUMNS } from './displayed-columns.const';
